Walić Mapke

+++ Tanks2/TileMap.h
@@ -8,7 +8,7 @@ public:
 	TileMap();
 	~TileMap();
 
-	bool load(const std::string& tileset, sf::Vector2i tileSize, const int* tiles, unsigned int width, unsigned int height);
+	bool load(const std::string& tileset, sf::Vector2f tileSize, const int* tiles, unsigned int width, unsigned int height);
 
 private:
 	virtual void draw(sf::RenderTarget& target, sf::RenderStates states) const;
